Renamed AR Classes. Fixed ARCameraActivity and ManagerAR dependencies.
